[bos #40598] solution to some test failing in ubuntu22.
Add PY_SSIZE_T_CLEAN definition to header files.

Small modif to fix StdAloneYacsLoaderTest1.py

+++ src/runtime/PyStdout.cxx
@@ -52,7 +52,7 @@ static PyObject *
 PyStdOut_write(PyStdOut *self, PyObject *args)
 {
   char *c;
-  int l;
+  Py_ssize_t l;
   if (!PyArg_ParseTuple(args, "s#:write",&c, &l))
     return NULL;
